updated to allow postgresql to instantiate on heroku

+++ db/postgresql_setup.py
@@ -1,30 +1,39 @@
 #This setup file creates a postgresql database 
-import config
-#print(config.test_msg)
-
 import psycopg2
+import os
 
-#establishing the connection
-conn = psycopg2.connect(
-   database="postgres", user='postgres', password=config.postgres_pwd, host='127.0.0.1', port= '5432'
-)
+table_name = "color_votes"
+
+if 'DATABASE_URL' in os.environ:
+   postgres_url = os.environ['DATABASE_URL']
+else:
+   #if we're not running in heroku then try and get my local config password
+   import config
+   table_name = config.table_name
+   postgres_url = f"postgresql://postgres:{config.postgres_pwd}@127.0.0.1:5432/{config.db_name}"
+
+   #if local, try and delete and recreate the database to start afresh
+   #establishing the connection
+   conn = psycopg2.connect(f"postgresql://postgres:{config.postgres_pwd}@127.0.0.1:5432/postgres")
+
+   #Creating a cursor object using the cursor() method
+   cursor = conn.cursor()
+
+   conn.autocommit = True
+
+   cursor.execute(f"DROP DATABASE IF EXISTS {config.db_name}")
+
+   #Creating a database
+   cursor.execute(f"CREATE DATABASE {config.db_name}")
+   print("Database created successfully........")
+
+   #close the connection to reestablish database connect to the newly create database
+   conn.close()
+
+conn = psycopg2.connect(postgres_url)
+
 conn.autocommit = True
 
-#Creating a cursor object using the cursor() method
-cursor = conn.cursor()
-
-cursor.execute(f"DROP DATABASE IF EXISTS {config.db_name}")
-
-#Creating a database
-cursor.execute(f"CREATE DATABASE {config.db_name}")
-print("Database created successfully........")
-
-#close the connection to reestablish database connect to the newly create database
-conn.close()
-
-conn = psycopg2.connect(
-   database=config.db_name, user='postgres', password=config.postgres_pwd, host='127.0.0.1', port= '5432'
-)
 cursor = conn.cursor()
 
 #define data
@@ -36,9 +45,9 @@ sp_to_en_colors = [{"votes": 62, "color":"red"},
                     {"votes": 64, "color":"black"},
                     {"votes": 48, "color":"pink"}]
 
-cursor.execute(f"DROP TABLE IF EXISTS {config.table_name}")
+cursor.execute(f"DROP TABLE IF EXISTS {table_name}")
 
-sql_query = f'''CREATE TABLE {config.table_name}(
+sql_query = f'''CREATE TABLE {table_name}(
    ID SERIAL PRIMARY KEY,
    VOTES        INT,
    COLOR        VARCHAR(50)
@@ -51,14 +60,14 @@ print("Table created successfully........")
 
 
 for sp_to_en_color in sp_to_en_colors:
-    cursor.execute(f'''INSERT INTO {config.table_name}(
+    cursor.execute(f'''INSERT INTO {table_name}(
     VOTES, COLOR) VALUES 
     ('{sp_to_en_color["votes"]}', '{sp_to_en_color["color"]}')''')
 
 conn.commit()
 print("Data added successfully........")
 
-cursor.execute(f'''SELECT * from {config.table_name}''')
+cursor.execute(f'''SELECT * from {table_name}''')
 
 results = cursor.fetchall()
 print(results)
